refactor: Single Subject Details
